pyproject.toml
[project]
name = "atlas-importer"
version = "0.1.0"
description = "CSV to Cloudera Atlas Business Glossary Importer"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
]
